Update with PR reviews
